fix(dns_record): prevent FQDN normalization from swallowing name shortening changes
The ModifyPlan FQDN drift suppression used a blind prefix check that
incorrectly treated genuine name changes as drift when the new name was
a prefix of the old FQDN (e.g. changing 'resend._domainkey.app' to
'resend._domainkey').

Fix: extract zone_name from the DNS record API response (already being
made in Create/Update/Read), store it in Terraform private state, and
use it in ModifyPlan to verify the suffix matches the zone exactly
before suppressing the diff.

Also moves all custom plan-modification logic from the codegen-managed
resource.go into custom.go to reduce merge conflicts.

Fixes: https://github.com/cloudflare/terraform-provider-cloudflare/issues/6566
Ref: APIX-725
